added access control

+++ src/main/java/com/gym/controller/InstructorController.java
@@ -1,17 +1,20 @@
 package com.gym.controller;
 
 import com.gym.aspect.TrackExecutionTime;
-import com.gym.domain.dto.ClientDTO;
-import com.gym.domain.dto.InstructorDTO;
-import com.gym.domain.dto.InstructorRequest;
-import com.gym.domain.dto.PlanDTO;
+import com.gym.domain.dto.*;
+import com.gym.domain.entity.Instructor;
+import com.gym.domain.entity.User;
 import com.gym.domain.exception.ResourceNotFoundException;
+import com.gym.domain.exception.SortingException;
 import com.gym.domain.filter.Filter;
 import com.gym.domain.mapper.InstructorConverter;
 import com.gym.service.InstructorService;
+import com.gym.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
 
@@ -24,6 +27,8 @@ import java.util.Objects;
 public class InstructorController {
     @Autowired
     private InstructorService instructorService;
+    @Autowired
+    private UserService userService;
 
     @GetMapping
     @TrackExecutionTime
@@ -55,12 +60,16 @@ public class InstructorController {
 
     @GetMapping("/{instructorId}")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_INSTRUCTOR')")
-    public ResponseEntity<InstructorDTO> findById(@PathVariable Integer instructorId){
-        InstructorDTO instructorDTO = instructorService.findById(instructorId);
-        if(instructorDTO == null){
-            throw new ResourceNotFoundException("Exercise not found.");
+    public ResponseEntity<InstructorDTO> findById(@PathVariable Integer instructorId, Authentication auth){
+        if(isAuthorized(auth, instructorId)){
+            InstructorDTO instructorDTO = instructorService.findById(instructorId);
+            if(instructorDTO == null){
+                throw new ResourceNotFoundException("Exercise not found.");
+            }
+            return ResponseEntity.ok(instructorDTO);
+        }else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
-        return ResponseEntity.ok(instructorDTO);
     }
 
     @PostMapping
@@ -77,25 +86,34 @@ public class InstructorController {
 
     @PutMapping("/{instructorId}")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_INSTRUCTOR')")
-    public ResponseEntity<Void> updateInstructor(@PathVariable Integer instructorId, @RequestBody InstructorRequest request){
-        InstructorDTO instructorDTO = InstructorConverter.toDTO(
-                InstructorConverter.toEntity(request)
-        );
-        instructorDTO.setId(instructorId);
-        InstructorDTO response = this.instructorService.update(instructorDTO);
-        return response!=null? ResponseEntity.noContent().build():
-                ResponseEntity.notFound().build();
+    public ResponseEntity<Void> updateInstructor(@PathVariable Integer instructorId, @RequestBody InstructorRequest request, Authentication auth){
+        if(isAuthorized(auth, instructorId)){
+            InstructorDTO instructorDTO = InstructorConverter.toDTO(
+                    InstructorConverter.toEntity(request)
+            );
+            instructorDTO.setId(instructorId);
+            InstructorDTO response = this.instructorService.update(instructorDTO);
+            return response!=null? ResponseEntity.noContent().build():
+                    ResponseEntity.notFound().build();
+        }else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
     }
     @DeleteMapping("/{instructorId}/delete")
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_INSTRUCTOR')")
-    public ResponseEntity<Void> deleteInstructor(@PathVariable Integer instructorId){
-        this.instructorService.delete(instructorService.findById(instructorId));
-        return  ResponseEntity.noContent().build();
+    public ResponseEntity<Void> deleteInstructor(@PathVariable Integer instructorId, Authentication auth){
+        if(isAuthorized(auth, instructorId)){
+            this.instructorService.delete(instructorService.findById(instructorId));
+            return  ResponseEntity.noContent().build();
+        }else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     //Get all plans for a given instructor.
     @GetMapping("/{instructorId}/plans")
-    @PreAuthorize("hasAnyRole('ROLE_CLIENT', 'ROLE_INSTRUCTOR')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_CLIENT', 'ROLE_INSTRUCTOR')")
     public ResponseEntity<List<PlanDTO>> getAllPlansForGivenInstructor(@PathVariable Integer instructorId,
                                                                        @RequestParam(required = false) String name,
                                                                        @RequestParam(required = false) String sort,
@@ -125,34 +143,49 @@ public class InstructorController {
     //Get all clients for a given instructor.
     @GetMapping("/{instructorId}/clients")
     @TrackExecutionTime
-    @PreAuthorize("hasRole('ROLE_INSTRUCTOR')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_INSTRUCTOR')")
     public ResponseEntity<List<ClientDTO>> getAllClientsForGivenInstructor(@PathVariable Integer instructorId,
                                                                            @RequestParam(required = false) String name,
                                                                            @RequestParam(required = false) String planId,
                                                                            @RequestParam(required = false) String sort,
                                                                            @RequestParam(required = false) Integer pageNumber,
-                                                                           @RequestParam(required = false) Integer pageSize){
-        if(name != null){
-            name = name.replace("%20"," ");
-        }
-        Filter nameFilter = new Filter("name", name, "LIKE", null,pageNumber,pageSize);
-        Filter planFilter = new Filter("id", planId, "=", null, pageNumber, pageSize);
-        if(sort != null){
-            String[] sortValue = sort.split(":");
-            if(Objects.equals(sortValue[0], "name")){
-                nameFilter.setSort(sortValue[1]);
-            } else if (Objects.equals(sortValue[0], "planId")) {
-                planFilter.setSort(sortValue[1]);
-            } else {
-                throw new RuntimeException("Invalid sort field.");
+                                                                           @RequestParam(required = false) Integer pageSize,
+                                                                           Authentication auth){
+        if(isAuthorized(auth, instructorId)){
+            if(name != null){
+                name = name.replace("%20"," ");
             }
-        }
+            Filter nameFilter = new Filter("name", name, "LIKE", null,pageNumber,pageSize);
+            Filter planFilter = new Filter("id", planId, "=", null, pageNumber, pageSize);
+            if(sort != null){
+                String[] sortValue = sort.split(":");
+                if(Objects.equals(sortValue[0], "name")){
+                    nameFilter.setSort(sortValue[1]);
+                } else if (Objects.equals(sortValue[0], "planId")) {
+                    planFilter.setSort(sortValue[1]);
+                } else {
+                    throw new SortingException("Invalid sort field.");
+                }
+            }
 
-        List<ClientDTO> clients = instructorService.findAllClientsByInstructorId(instructorId,nameFilter,planFilter);
-        if(clients == null){
-            throw new ResourceNotFoundException("Clients not found.");
+            List<ClientDTO> clients = instructorService.findAllClientsByInstructorId(instructorId,nameFilter,planFilter);
+            if(clients == null){
+                throw new ResourceNotFoundException("Clients not found.");
+            }
+            return ResponseEntity.ok(clients);
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
-        return ResponseEntity.ok(clients);
+    }
+
+    private boolean isAuthorized(Authentication authentication, Integer instructorId) {
+        String authenticatedUsername = authentication.getName();
+        if(authenticatedUsername.equals("admin")){
+            return true;
+        }
+        User user = (User) userService.loadUserByUsername(authenticatedUsername);
+        InstructorDTO instructorDTO = instructorService.findById(user.getInstructor().getId());
+        return Objects.equals(instructorDTO.getId(), instructorId);
     }
 
 }
